Create service to inject script
